requirements.txt
transformers
torch
accelerate
gradio
openai-whisper
soundfile